comparing controllers

+++ src/main/java/com/example/clientConnect/client/Client.java
@@ -99,7 +99,7 @@ public class Client {
         return balance;
     }
 
-    public void setBalance(Integer balance) {
+    public void setBalance(double balance) {
         this.balance = balance;
     }
